Switched to Float to contain SensorReading value

+++ src/main/java/net/kf03w5t5741l/sensorbase/server/api/SensorReadingEndpoints.java
@@ -42,7 +42,7 @@ public class SensorReadingEndpoints {
 
     @GetMapping("/gteq/{value}")
     public List<SensorReading> getSensorReadingsGtEq(
-            @PathVariable Integer value) {
+            @PathVariable Float value) {
         return this.sensorReadingService.findValueGtEq(value);
     }
